requirements.txt
pytest
numpy==1.19.2
fastapi==0.70.0
typing-extensions==3.7.4.3
pydantic==1.8.2
pillow==8.4.0
tensorflow==2.6.1
opencv-python-headless==4.5.4.58
gunicorn==20.1.0
uvicorn==0.15.0
pypyr==4.6.0
boto3==1.19.8
python-dotenv==0.19.1
python-multipart==0.0.5
pre-commit